Added getRecordAsXml that returns an record as XML.
In order for this to work specify the xml option in a record. 
The xml option is an array that has 2 keys. 

ignore_fields is an array of fields that are not displayed in the xml
include_relations(optional) is an array of relations that should be included in
the xml. 



git-svn-id: file:///var/www/production/svnmirror/svn_doctrine/trunk/lib@1281 2b2ae520-0f61-44a6-8c59-7e116ffd6265