[1.3, 1.4] Fixed setting of Doctrine's default culture before records are initialized
Before this patch, sfDoctrineRecord::listenToChangeCultureEvent() was only registered
once the first model was initialized. If the event was fired before that, the internal
property $_defaultCulture was not updated.

After this patch, the listener is already registered in the plugin configuration.
The listener is guaranteed to be called on each request, because the constructor of
sfUser calls sfUser::setCulture() implicitely, which fires the event.


git-svn-id: file:///var/www/production/sfweb/svn/branches/1.4@29156 ee427ae8-e902-0410-961c-c3ed070cd9f9