Fix for DC-797: one-to-one or none causing dirty hydration
This patch modifies the record's _isValueModified method to mark a
change from null to Doctrine_Null (or vice versa) as not a modification.

This bug causes any record containing an empty one-to-one relations
(nullable foreign key) to be hydrated as dirty off-the-bat, when no
modifications have taken place. This causes knock-on problems with
things like the Timestampable behavior.

This bug is discussed at doctrine/doctrine1#8